requirements.txt
Jinja2>=3.1.2
PyYAML>=6.0 
